requirements.txt
Django==5.0.6
djangorestframework==3.15.1
tzdata==2024.1
python-decouple==3.8
Pillow==10.3.0
psycopg==3.1.19
psycopg-binary==3.1.19
